+++ src/Fable.Transforms/Replacements.fs
@@ -293,8 +293,8 @@ let coreModFor = function
     | FSharpResult _ -> "Option"
     | FSharpChoice _ -> "Option"
     | FSharpReference _ -> "Types"
-    | BclHashSet _
-    | BclDictionary _
+    | BclHashSet _ -> "MutableSet"
+    | BclDictionary _ -> "MutableMap"
     | BclKeyValuePair _ -> failwith "Cannot decide core module"
 
 let makeLongInt r t signed (x: uint64) =
@@ -1683,8 +1683,8 @@ let resizeArrays (com: ICompiler) (ctx: Context) r (t: Type) (i: CallInfo) (this
         Helper.GlobalCall("Array", t, args, memb="from", ?loc=r) |> Some
     | "get_Item", Some ar, [idx] -> getExpr r t ar idx |> Some
     | "set_Item", Some ar, [idx; value] -> Set(ar, ExprSet idx, value, r) |> Some
-    | "Add", Some ar, args ->
-        Helper.InstanceCall(ar, "push", t, args, ?loc=r) |> Some
+    | "Add", Some ar, [arg] ->
+        Helper.CoreCall("Array", "addInPlace", t, [arg; ar], ?loc=r) |> Some
     | "Remove", Some ar, [arg] ->
         Helper.CoreCall("Array", "removeInPlace", t, [arg; ar], ?loc=r) |> Some
     | "RemoveAll", Some ar, [arg] ->
@@ -1904,15 +1904,11 @@ let optionModule (com: ICompiler) (ctx: Context) r (t: Type) (i: CallInfo) (_: E
     | "GetValue", [c] -> Get(c, OptionValue, t, r) |> Some
     | ("OfObj" | "OfNullable"), [c] -> TypeCast(c, t) |> Some
     | ("ToObj" | "ToNullable" | "Flatten"), [c] ->
-        Helper.CoreCall("Option", "value", t, [c; makeBoolConst true], ?loc=r) |> Some
+        Helper.CoreCall("Option", "tryValue", t, args, ?loc=r) |> Some
     | "IsSome", [c] -> Test(c, OptionTest true, r) |> Some
     | "IsNone", [c] -> Test(c, OptionTest false, r) |> Some
-    | ("Map" | "Bind" as meth), args ->
+    | ("Filter" | "Map" | "Map2" | "Map3" | "Bind" as meth), args ->
         Helper.CoreCall("Option", Naming.lowerFirst meth, t, args, i.SignatureArgTypes, ?loc=r) |> Some
-    | ("Map2" | "Map3"), args ->
-        Helper.CoreCall("Option", "mapMultiple", t, args, i.SignatureArgTypes, ?loc=r) |> Some
-    | "Filter", _ ->
-        Helper.CoreCall("Option", "filter", t, args, i.SignatureArgTypes, ?loc=r) |> Some
     | "ToArray", [arg] ->
         toArray r t arg |> Some
     | "FoldBack", [folder; opt; state] ->
@@ -2048,6 +2044,9 @@ let bigints (com: ICompiler) (ctx: Context) r (t: Type) (i: CallInfo) (thisArg: 
             | Decimal -> toDecimal com ctx r t args |> Some
             | BigInt -> None
         | _ -> None
+    | None, "DivRem" ->
+        let args = List.take 2 args // implementation takes 2 args, ignore the third arg
+        Helper.CoreCall("BigInt", "divRem", t, args, i.SignatureArgTypes, ?loc=r) |> Some
     | None, meth when meth.StartsWith("get_") ->
         Helper.CoreValue("BigInt", meth, t) |> Some
     | callee, meth ->
@@ -2067,6 +2066,13 @@ let languagePrimitives (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisAr
     | "DivideByInt", _ -> applyOp com ctx r t i.CompiledName args i.SignatureArgTypes i.GenericArgs |> Some
     | "GenericZero", _ -> getZero com ctx t |> Some
     | "GenericOne", _ -> getOne com ctx t |> Some
+    | ("SByteWithMeasure"
+    | "Int16WithMeasure"
+    | "Int32WithMeasure"
+    | "Int64WithMeasure"
+    | "Float32WithMeasure"
+    | "FloatWithMeasure"
+    | "DecimalWithMeasure"), [arg] -> arg |> Some
     | "EnumOfValue", [arg] ->
         match t with
         | Enum e -> EnumConstant(arg, e) |> makeValue r |> Some
@@ -2332,23 +2338,28 @@ let log (_: ICompiler) r t (i: CallInfo) (_: Expr option) (args: Expr list) =
         | _ -> [args.Head]
     Helper.GlobalCall("console", t, args, memb="log", ?loc=r)
 
-let bitConvert (_: ICompiler) (ctx: Context) r (_: Type) (i: CallInfo) (_: Expr option) (args: Expr list) =
-    let memberName =
-        if i.CompiledName = "GetBytes" then
+let bitConvert (com: ICompiler) (ctx: Context) r t (i: CallInfo) (_: Expr option) (args: Expr list) =
+    match i.CompiledName with
+    | "GetBytes" ->
+        let memberName =
             match args.Head.Type with
             | Boolean -> "getBytesBoolean"
             | Char | String -> "getBytesChar"
             | Number Int16 -> "getBytesInt16"
             | Number Int32 -> "getBytesInt32"
-            | Builtin BclInt64 -> "getBytesInt64"
             | Number UInt16 -> "getBytesUInt16"
-            | Builtin BclUInt64 -> "getBytesUInt64"
             | Number UInt32 -> "getBytesUInt32"
             | Number Float32 -> "getBytesSingle"
             | Number Float64 -> "getBytesDouble"
+            | Builtin BclInt64 -> "getBytesInt64"
+            | Builtin BclUInt64 -> "getBytesUInt64"
             | x -> failwithf "Unsupported type in BitConverter.GetBytes(): %A" x
-        else Naming.lowerFirst i.CompiledName
-    Helper.CoreCall("BitConverter", memberName, Boolean, args, i.SignatureArgTypes, ?loc=r) |> Some
+        let expr = Helper.CoreCall("BitConverter", memberName, Boolean, args, i.SignatureArgTypes, ?loc=r)
+        if com.Options.typedArrays then expr |> Some
+        else toArray com t expr |> Some // convert to dynamic array
+    | _ ->
+        let memberName = Naming.lowerFirst i.CompiledName
+        Helper.CoreCall("BitConverter", memberName, Boolean, args, i.SignatureArgTypes, ?loc=r) |> Some
 
 let convert (com: ICompiler) (ctx: Context) r t (i: CallInfo) (_: Expr option) (args: Expr list) =
     match i.CompiledName with
@@ -2587,11 +2598,16 @@ let regex com (ctx: Context) r t (i: CallInfo) (thisArg: Expr option) (args: Exp
         let meth = Naming.removeGetSetPrefix meth |> Naming.lowerFirst
         Helper.CoreCall("RegExp", meth, t, args, i.SignatureArgTypes, ?thisArg=thisArg, ?loc=r) |> Some
 
-let encoding (_: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr option) (args: Expr list) =
+let encoding (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr option) (args: Expr list) =
     match i.CompiledName, thisArg, args.Length with
     | ("get_Unicode" | "get_UTF8"), _, _ ->
         Helper.CoreCall("Encoding", i.CompiledName, t, args, i.SignatureArgTypes, ?loc=r) |> Some
-    | ("GetBytes" | "GetString"), Some callee, (1 | 3) ->
+    | "GetBytes", Some callee, (1 | 3) ->
+        let meth = Naming.lowerFirst i.CompiledName
+        let expr = Helper.InstanceCall(callee, meth, t, args, i.SignatureArgTypes, ?loc=r)
+        if com.Options.typedArrays then expr |> Some
+        else toArray com t expr |> Some // convert to dynamic array
+    | "GetString", Some callee, (1 | 3) ->
         let meth = Naming.lowerFirst i.CompiledName
         Helper.InstanceCall(callee, meth, t, args, i.SignatureArgTypes, ?loc=r) |> Some
     | _ -> None