+++ src/fable-library/Array.fs
@@ -202,35 +202,35 @@ let mapIndexed3 (f: int->'T1->'T2->'T3->'U) (source1: 'T1[]) (source2: 'T2[]) (s
         result.[i] <- f i source1.[i] source2.[i] source3.[i]
     result
 
-let map3 f (source1: 'T[]) (source2: 'U[]) (source3: 'U[]) ([<Inject>] cons: IArrayCons<'W>): 'W[] =
+let map3 (f: 'T1->'T2->'T3->'U) (source1: 'T1[]) (source2: 'T2[]) (source3: 'T3[]) ([<Inject>] cons: IArrayCons<'U>): 'U[] =
     if source1.Length <> source2.Length || source2.Length <> source3.Length then failwith "Arrays had different lengths"
     let result = cons.Create(source1.Length)
     for i = 0 to source1.Length - 1 do
         result.[i] <- f source1.[i] source2.[i] source3.[i]
     result
 
-let mapFold<'T,'State,'Result> (mapping: 'State -> 'T -> 'Result * 'State) state (array: 'T[]) ([<Inject>] cons: IArrayCons<'Result>) =
+let mapFold<'T, 'State, 'Result> (mapping: 'State -> 'T -> 'Result * 'State) state (array: 'T[]) ([<Inject>] cons: IArrayCons<'Result>) =
     match array.Length with
     | 0 -> [| |], state
     | len ->
         let mutable acc = state
         let res = cons.Create len
         for i = 0 to array.Length-1 do
-            let h',s' = mapping acc array.[i]
-            res.[i] <- h'
-            acc <- s'
+            let h,s = mapping acc array.[i]
+            res.[i] <- h
+            acc <- s
         res, acc
 
-let mapFoldBack<'T,'State,'Result> (mapping: 'T -> 'State -> 'Result * 'State) (array: 'T[]) state ([<Inject>] cons: IArrayCons<'Result>) =
+let mapFoldBack<'T, 'State, 'Result> (mapping: 'T -> 'State -> 'Result * 'State) (array: 'T[]) state ([<Inject>] cons: IArrayCons<'Result>) =
     match array.Length with
     | 0 -> [| |], state
     | len ->
         let mutable acc = state
         let res = cons.Create len
         for i = array.Length-1 downto 0 do
-            let h',s' = mapping array.[i] acc
-            res.[i] <- h'
-            acc <- s'
+            let h,s = mapping array.[i] acc
+            res.[i] <- h
+            acc <- s
         res, acc
 
 let indexed (source: 'T[]) =
@@ -273,7 +273,7 @@ let collect (mapping: 'T -> 'U[]) (array: 'T[]) ([<Inject>] cons: IArrayCons<'U>
 
 let countBy (projection: 'T -> 'Key) (array: 'T[]) ([<Inject>] eq: IEqualityComparer<'Key>) =
     let dict = Dictionary<'Key, int>(eq)
-    let keys = [||]
+    let keys: 'Key[] = [||]
     for value in array do
         let key = projection value
         match dict.TryGetValue(key) with
@@ -313,7 +313,7 @@ let except (itemsToExclude: seq<'T>) (array: 'T[]) ([<Inject>] eq: IEqualityComp
 
 let groupBy (projection: 'T -> 'Key) (array: 'T[]) ([<Inject>] cons: IArrayCons<'T>) ([<Inject>] eq: IEqualityComparer<'Key>) =
     let dict = Dictionary<'Key, 'T list>(eq)
-    let keys = [||]
+    let keys: 'Key[] = [||]
     for v in array do
         let key = projection v
         match dict.TryGetValue(key) with
@@ -348,7 +348,7 @@ let pairwise (array: 'T[]) =
         let count = array.Length - 1
         let result = newDynamicArrayImpl count
         for i = 0 to count - 1 do
-            result.[i] <-  array.[i], array.[i+1]
+            result.[i] <- array.[i], array.[i+1]
         result
 
 let replicate count initial ([<Inject>] cons: IArrayCons<'T>) =
@@ -426,6 +426,10 @@ let takeWhile predicate (array: 'T[]) ([<Inject>] cons: IArrayCons<'T>) =
         emptyImpl cons
     else
         subArrayImpl array 0 count
+
+let addInPlace (x: 'T) (array: 'T[]) =
+    // if isTypedArrayImpl array then invalidArg "array" "Typed arrays not supported"
+    pushImpl array x |> ignore
 
 let addRangeInPlace (range: seq<'T>) (array: 'T[]) =
     // if isTypedArrayImpl array then invalidArg "array" "Typed arrays not supported"
@@ -601,14 +605,14 @@ let forAll predicate (array: 'T[]) =
 let permute f (array: 'T[]) =
     let size = array.Length
     let res = copyImpl array
-    let checkFlags = newDynamicArrayImpl size
+    let checkFlags = Array.zeroCreate size
     iterateIndexed (fun i x ->
         let j = f i
         if j < 0 || j >= size then
             invalidOp "Not a valid permutation"
         res.[j] <- x
         checkFlags.[j] <- 1) array
-    let isValid = forAll ((=) 1) checkFlags
+    let isValid = checkFlags |> forAllImpl ((=) 1)
     if not isValid then
         invalidOp "Not a valid permutation"
     res
@@ -649,14 +653,14 @@ let sortByDescending (projection: 'a->'b) (xs: 'a[]) ([<Inject>] comparer: IComp
 let sortWith (comparer: 'T -> 'T -> int) (xs: 'T[]): 'T[] =
     sortInPlaceWith comparer (copyImpl xs)
 
-let unfold<'T,'State> (generator: 'State -> ('T*'State) option) (state: 'State): 'State[] =
-    let res = [||]
+let unfold<'T, 'State> (generator: 'State -> ('T*'State) option) (state: 'State): 'T[] =
+    let res: 'T[] = [||]
     let rec loop state =
         match generator state with
         | None -> ()
-        | Some (x,s') ->
+        | Some (x, s) ->
             pushImpl res x |> ignore
-            loop s'
+            loop s
     loop state
     res
 
@@ -761,7 +765,7 @@ let tryItem index (array: 'T[]) =
     if index < 0 || index >= array.Length then None
     else Some array.[index]
 
-let foldBackIndexed<'T,'State> folder (array: 'T[]) (state: 'State) =
+let foldBackIndexed<'T, 'State> folder (array: 'T[]) (state: 'State) =
     // if isTypedArrayImpl array then
     //     let mutable acc = state
     //     let size = array.Length
@@ -771,7 +775,7 @@ let foldBackIndexed<'T,'State> folder (array: 'T[]) (state: 'State) =
     // else
     foldBackIndexedImpl (fun acc x i -> folder i x acc) state array
 
-let foldBack<'T,'State> folder (array: 'T[]) (state: 'State) =
+let foldBack<'T, 'State> folder (array: 'T[]) (state: 'State) =
     // if isTypedArrayImpl array then
     //     foldBackIndexed (fun _ x acc -> folder x acc) array state
     // else
@@ -887,7 +891,7 @@ let toList (source: 'T[]) =
 let windowed (windowSize: int) (source: 'T[]): 'T[][] =
     if windowSize <= 0 then
         failwith "windowSize must be positive"
-    let res = FSharp.Core.Operators.max 0 (source.Length - windowSize) |> Helpers.newDynamicArrayImpl
+    let res = FSharp.Core.Operators.max 0 (source.Length - windowSize) |> newDynamicArrayImpl
     for i = windowSize to source.Length do
         res.[i - windowSize] <- source.[i-windowSize..i-1]
     res