+++ src/fable-library/BitConverter.ts
@@ -1,3 +1,4 @@
+import { uint8 } from "./Int32";
 import Long, { fromBits, getHighBits, getHighBitsUnsigned, getLowBits, getLowBitsUnsigned } from "./Long";
 
 const littleEndian = true;
@@ -95,67 +96,78 @@ export function doubleToInt64Bits(value: number) {
   return fromBits(lowBits, highBits, false);
 }
 
-export function toBoolean(bytes: Uint8Array, offset: number): boolean {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toBoolean(bytes: ArrayLike<uint8>, offset: number): boolean {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   return view.getUint8(offset) === 1 ? true : false;
 }
 
-export function toChar(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toChar(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   const code = view.getUint16(offset, littleEndian);
   return String.fromCharCode(code);
 }
 
-export function toInt16(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toInt16(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   return view.getInt16(offset, littleEndian);
 }
 
-export function toInt32(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toInt32(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   return view.getInt32(offset, littleEndian);
 }
 
-export function toInt64(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toInt64(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   const lowBits = view.getInt32(offset + (littleEndian ? 0 : 4), littleEndian);
   const highBits = view.getInt32(offset + (littleEndian ? 4 : 0), littleEndian);
   return fromBits(lowBits, highBits, false);
 }
 
-export function toUInt16(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toUInt16(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   return view.getUint16(offset, littleEndian);
 }
 
-export function toUInt32(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toUInt32(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   return view.getUint32(offset, littleEndian);
 }
 
-export function toUInt64(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toUInt64(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   const lowBits = view.getUint32(offset + (littleEndian ? 0 : 4), littleEndian);
   const highBits = view.getUint32(offset + (littleEndian ? 4 : 0), littleEndian);
   return fromBits(lowBits, highBits, true);
 }
 
-export function toSingle(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toSingle(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   return view.getFloat32(offset, littleEndian);
 }
 
-export function toDouble(bytes: Uint8Array, offset: number) {
-  const view = new DataView(bytes.buffer, bytes.byteOffset, bytes.byteLength);
+export function toDouble(bytes: ArrayLike<uint8>, offset: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  const view = new DataView(array.buffer, array.byteOffset, array.byteLength);
   return view.getFloat64(offset, littleEndian);
 }
 
-export function toString(bytes: Uint8Array, offset?: number, count?: number) {
-  let ar = bytes;
-  if (typeof offset !== "undefined" && typeof count !== "undefined") {
-    ar = bytes.subarray(offset, offset + count);
-  } else if (typeof offset !== "undefined") {
-    ar = bytes.subarray(offset);
+export function toString(bytes: ArrayLike<uint8>, offset?: number, count?: number) {
+  const array = ArrayBuffer.isView(bytes) ? bytes : Uint8Array.from(bytes);
+  let buffer = Buffer.from(array.buffer, array.byteOffset, array.byteLength);
+  if (offset != null && count != null) {
+    buffer = buffer.subarray(offset, offset + count);
+  } else if (offset != null) {
+    buffer = buffer.subarray(offset);
   }
-  return Array.from(ar).map((b) => ("0" + b.toString(16)).slice(-2)).join("-");
+  return Array.from(buffer).map((b) => ("0" + b.toString(16)).slice(-2)).join("-");
 }