+++ src/fable-library/Option.ts
@@ -10,13 +10,13 @@ import { compare, equals, structuralHash } from "./Util";
 // 2- To get the value of an option the `getValue` helper
 //    below must **always** be used.
 
-export type Option<T> = T | Some<T> | null;
+export type Option<T> = T | Some<T> | null | undefined;
 
 // Using a class here for better compatibility with TS files importing Some
 export class Some<T> {
-  public value: T | null;
+  public value: T;
 
-  constructor(value: T | null) {
+  constructor(value: T) {
     this.value = value;
   }
 
@@ -50,44 +50,52 @@ export class Some<T> {
   }
 }
 
-export function some<T>(x: T | null): Option<T> {
-  x = (x === undefined) ? null : x;
+export function some<T>(x: T): Option<T> {
   return x == null || x instanceof Some ? new Some(x) : x;
 }
 
-export function value<T>(x: Option<T>, acceptNull?: boolean) {
+export function value<T>(x: Option<T>) {
   if (x == null) {
-    if (!acceptNull) {
-      throw new Error("Option has no value");
-    }
-    return null;
+    throw new Error("Option has no value");
   } else {
     return x instanceof Some ? x.value : x;
   }
 }
 
-export function defaultArg<T>(arg: Option<T>, defaultValue: T, f?: (arg0: T | null) => T) {
-  return arg == null ? defaultValue : (f != null ? f(value(arg)) : value(arg));
+export function tryValue<T>(x: Option<T>) {
+  return x instanceof Some ? x.value : x;
 }
 
-export function defaultArgWith<T>(arg: Option<T>, defThunk: () => T) {
-  return arg == null ? defThunk() : value(arg);
+export function defaultArg<T>(opt: Option<T>, defaultValue: T, f?: (arg: T) => T) {
+  return (opt != null) ? (f != null ? f(value(opt)) : value(opt)) : defaultValue;
 }
 
-export function filter<T>(predicate: (arg0: T | null) => boolean, arg: Option<T>) {
-  return arg != null ? (!predicate(value(arg)) ? null : arg) : arg;
+export function defaultArgWith<T>(opt: Option<T>, defThunk: () => T) {
+  return (opt != null) ? value(opt) : defThunk();
 }
 
-export function map<T, U>(f: (arg0: T | null) => U, arg: Option<T>): Option<U> {
-  return arg == null ? null : some(f(value(arg)));
+export function filter<T>(predicate: (arg: T) => boolean, opt: Option<T>) {
+  return (opt != null) ? (predicate(value(opt)) ? opt : null) : opt;
 }
 
-export function mapMultiple<T>(predicate: (...args: T[]) => boolean, ...args: T[]) {
-  return args.every((x) => x != null) ? predicate.apply(null, args) : null;
+export function map<T, U>(mapping: (arg: T) => U, opt: Option<T>): Option<U> {
+  return (opt != null) ? some(mapping(value(opt))) : null;
 }
 
-export function bind<T, U>(f: (arg0: T | null) => Option<U>, arg: Option<T>): Option<U> {
-  return arg == null ? null : f(value(arg));
+export function map2<T1, T2, U>(
+  mapping: (arg1: T1, arg2: T2) => Option<U>,
+  opt1: Option<T1>, opt2: Option<T2>) {
+  return (opt1 != null && opt2 != null) ? mapping(value(opt1), value(opt2)) : null;
+}
+
+export function map3<T1, T2, T3, U>(
+  mapping: (arg1: T1, arg2: T2, arg3: T3) => Option<U>,
+  opt1: Option<T1>, opt2: Option<T2>, opt3: Option<T3>) {
+  return (opt1 != null && opt2 != null && opt3 != null) ? mapping(value(opt1), value(opt2), value(opt3)) : null;
+}
+
+export function bind<T, U>(binder: (arg: T) => Option<U>, opt: Option<T>): Option<U> {
+  return opt != null ? binder(value(opt)) : null;
 }
 
 export function tryOp<T, U>(op: (x: T) => U, arg: T): Option<U> {