+++ tests/Main/ApplicativeTests.fs
@@ -1105,6 +1105,24 @@ let tests7 = [
         equal 3 y
 ]
 
+// Test ported directly from https://github.com/fable-compiler/Fable/pull/1336/files
+let private SomeProcedure (s: string): unit = failwith "Never called"
+let private SomeFunction (s: string): string = s
+
+[<Fable.Core.Emit("$0.name")>]
+let private getName (f: obj) = failwith "JS only"
+
+let tests8 = 
+    [
+        #if FABLE_COMPILER
+        testCase "Functions passed as parameters don't generate intermediate anonymous functions" <| fun () ->
+            let functionName = getName SomeFunction
+            equal "SomeFunction" functionName
+            let procedureName = getName SomeProcedure
+            equal "SomeProcedure" procedureName
+        #endif
+    ]
+
 let tests =
     testList "Applicative" (
         tests1
@@ -1114,5 +1132,6 @@ let tests =
         @ tests5
         @ tests6
         @ tests7
+        @ tests8
         @ CurriedApplicative.tests
     )